requirements.txt
tensorflow==1.12.2
bert-tensorflow==1.0.1
numpy==1.16.4
transformers
